+++ app/components/list-item/asset-stock-item.js
@@ -1,4 +1,35 @@
 import Ember from 'ember';
 
-export default Ember.Component.extend({
+const { Component, computed } = Ember;
+
+export default Component.extend({
+  classNames: [ 'item' ],
+
+  didReceiveAttrs () {
+    this.set('details', this.get('showAttributes'));
+  },
+
+  attributes: computed('asset.customFields', function () {
+    let fields = this.get('asset.customFields'),
+        attributes = Ember.A();
+
+    for ( let key in fields ) {
+      if ( !fields.hasOwnProperty(key) ) {
+        continue;
+      }
+
+      attributes.pushObject({
+        key,
+        value: fields[key]
+      });
+    }
+
+    return attributes;
+  }),
+
+  actions: {
+    toggleProperty ( prop ) {
+      this.toggleProperty(prop);
+    }
+  }
 });